requirements/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements/requirements.in
#
amqp==5.1.1
    # via kombu
asgiref==3.5.2
    # via django
async-timeout==4.0.2
    # via redis
attrs==22.1.0
    # via jsonschema
billiard==3.6.4.0
    # via celery
boto3[crt]==1.24.73
    # via -r requirements/requirements.in
botocore==1.27.73
    # via
    #   boto3
    #   s3transfer
celery[redis]==5.2.7
    # via
    #   -r requirements/requirements.in
    #   django-celery
celery-progress==0.1.2
    # via -r requirements/requirements.in
certifi==2022.9.14
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
cryptography==38.0.1
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
deprecated==1.2.13
    # via redis
dj-stripe==2.6.2
    # via -r requirements/requirements.in
django==4.1.1
    # via
    #   -r requirements/requirements.in
    #   dj-stripe
    #   django-allauth
    #   django-anymail
    #   django-appconf
    #   django-storages
    #   djangorestframework
    #   drf-spectacular
    #   jsonfield
django-allauth==0.51.0
    # via -r requirements/requirements.in
django-anymail[mailgun]==8.6
    # via -r requirements/requirements.in
django-appconf==1.0.5
    # via django-queued-storage
django-celery==3.1.17
    # via django-queued-storage
django-cleanup==6.0.0
    # via -r requirements/requirements.in
django-environ==0.9.0
    # via -r requirements/requirements.in
django-queued-storage==0.8.0
    # via -r requirements/requirements.in
django-storages==1.13.1
    # via -r requirements/requirements.in
django-widget-tweaks==1.4.12
    # via -r requirements/requirements.in
djangorestframework==3.13.1
    # via
    #   -r requirements/requirements.in
    #   drf-spectacular
djangorestframework-api-key==2.2.0
    # via -r requirements/requirements.in
drf-spectacular==0.24.0
    # via -r requirements/requirements.in
hiredis==2.0.0
    # via redis
idna==3.3
    # via requests
inflection==0.5.1
    # via drf-spectacular
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonfield==3.1.0
    # via dj-stripe
jsonschema==4.16.0
    # via drf-spectacular
kombu==5.2.4
    # via celery
oauthlib==3.2.1
    # via requests-oauthlib
packaging==21.3
    # via redis
prompt-toolkit==3.0.31
    # via click-repl
psycopg2-binary==2.9.3
    # via -r requirements/requirements.in
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.4.0
    # via django-allauth
pyparsing==3.0.9
    # via packaging
pyrsistent==0.18.1
    # via jsonschema
python-dateutil==2.8.2
    # via botocore
python3-openid==3.2.0
    # via django-allauth
pytz==2022.2.1
    # via
    #   celery
    #   djangorestframework
pyyaml==6.0
    # via drf-spectacular
redis[hiredis]==4.3.4
    # via
    #   -r requirements/requirements.in
    #   celery
requests==2.28.1
    # via
    #   django-allauth
    #   django-anymail
    #   requests-oauthlib
    #   stripe
requests-oauthlib==1.3.1
    # via django-allauth
s3transfer==0.6.0
    # via boto3
six==1.16.0
    # via
    #   click-repl
    #   django-queued-storage
    #   python-dateutil
sqlparse==0.4.2
    # via django
stripe==4.1.0
    # via
    #   -r requirements/requirements.in
    #   dj-stripe
uritemplate==4.1.1
    # via drf-spectacular
urllib3==1.26.12
    # via
    #   botocore
    #   requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit
wrapt==1.14.1
    # via deprecated
